+++ test/setup.js
@@ -86,6 +86,24 @@
 					'Group 2': {label: 'Group 2', val: 'Group 2', $order: 4}
 				});
 			});
+			it('should render optgroups with duplicated options correctly', function(done) {
+				var test = setup_test(['<select>',
+					'<optgroup label="Group 1">',
+					'<option value="a">Item A</option>',
+					'<option value="b">Item B</option>',
+					'</optgroup>',
+					'<optgroup label="Group 2">',
+					'<option value="a">Item A</option>',
+					'<option value="b">Item B</option>',
+					'</optgroup>',
+					'</select>'].join(''), {});
+				test.selectize.refreshOptions(true);
+				window.setTimeout(function() {
+					assert.equal(test.selectize.$dropdown_content.find('.optgroup').length, 2, 'expect 2 optgroups');
+					assert.equal(test.selectize.$dropdown_content.find('.option').length, 4, 'expect 4 options');
+					done();
+				}, 0);
+			});
 			it('should add options in text form (no html entities)', function() {
 				var test = setup_test('<select><option selected value="a">&lt;hi&gt;</option></select>', {});
 				expect(test.selectize.options['a'].text).to.be.equal('<hi>');